misc: split auth tests (#348)
