+++ tests/ami/auth/test_login_callback.py
@@ -1,6 +1,4 @@
 import datetime
-import json
-from base64 import urlsafe_b64encode
 from typing import Any
 
 import pytest
@@ -11,77 +9,9 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app import env
-from app.auth import generate_nonce, jwt_cookie_auth
 from app.models import Nonce, ScheduledNotification, User
 from app.utils import build_fc_hash
-from tests.ami.utils import assert_query_fails_without_auth, login
 from tests.utils import url_contains_param
-
-
-async def test_generate_nonce(monkeypatch: pytest.MonkeyPatch) -> None:
-    FAKE_TIME = datetime.datetime(2020, 12, 25, 17, 5, 55)
-    FAKE_UUID = "fake-uuid"
-
-    class mock_datetime(datetime.datetime):
-        @classmethod
-        def now(cls, tz: datetime.tzinfo | None = datetime.timezone.utc):
-            return FAKE_TIME
-
-    monkeypatch.setattr("app.auth.uuid4", lambda: FAKE_UUID)
-    monkeypatch.setattr("app.auth.datetime.datetime", mock_datetime)
-    assert generate_nonce() == urlsafe_b64encode(f"{FAKE_UUID}-{FAKE_TIME}".encode("utf8")).decode(
-        "utf8"
-    )
-
-
-async def test_login_france_connect(
-    test_client: TestClient[Litestar],
-    db_session: AsyncSession,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    FAKE_NONCE = "some-random-nonce"
-    monkeypatch.setattr("app.controllers.auth.generate_nonce", lambda: FAKE_NONCE)
-    response = test_client.get("/login-france-connect", follow_redirects=False)
-    redirected_url = response.headers["location"]
-    all_nonces = (await db_session.execute(select(Nonce))).scalars().all()
-    assert len(all_nonces) == 1
-    nonce = all_nonces[0]
-    assert nonce.nonce == FAKE_NONCE
-    assert response.status_code == 302
-    redirected_url = response.headers["location"]
-    assert redirected_url.startswith(
-        f"{env.PUBLIC_FC_BASE_URL}{env.PUBLIC_FC_AUTHORIZATION_ENDPOINT}"
-    )
-    assert url_contains_param(
-        "scope",
-        "openid identite_pivot preferred_username email",
-        redirected_url,
-    )
-    assert url_contains_param(
-        "redirect_uri", env.PUBLIC_FC_PROXY or env.PUBLIC_FC_AMI_REDIRECT_URL, redirected_url
-    )
-    assert url_contains_param("response_type", "code", redirected_url)
-    assert url_contains_param("client_id", env.PUBLIC_FC_AMI_CLIENT_ID, redirected_url)
-    assert url_contains_param(
-        "state", f"{env.PUBLIC_FC_AMI_REDIRECT_URL}?state={nonce.id}", redirected_url
-    )
-    assert url_contains_param("nonce", nonce.nonce, redirected_url)
-    assert url_contains_param("acr_values", "eidas1", redirected_url)
-    assert url_contains_param("prompt", "login", redirected_url)
-    assert url_contains_param("idp_hint", env.PUBLIC_FC_AMI_IDP_HINT, redirected_url)
-
-
-async def test_login_france_connect_error(
-    test_client: TestClient[Litestar],
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    def fake_nonce():
-        raise Exception()
-
-    monkeypatch.setattr("app.controllers.auth.generate_nonce", fake_nonce)
-    response = test_client.get("/login-france-connect", follow_redirects=False)
-    redirected_url = response.headers["location"]
-    assert redirected_url == "https://localhost:5173/#/technical-error"
 
 
 async def test_login_callback(
@@ -618,395 +548,3 @@ async def test_login_callback_bad_token_info(
     assert url_contains_param("code", "missing_access_token", redirected_url)
     all_nonces = (await db_session.execute(select(Nonce))).scalars().all()
     assert len(all_nonces) == 0
-
-
-async def test_login_callback_address(
-    test_client: TestClient[Litestar],
-    db_session: AsyncSession,
-    httpx_mock: HTTPXMock,
-    monkeypatch: pytest.MonkeyPatch,
-    userinfo: dict[str, Any],
-    decoded_id_token: dict[str, Any],
-) -> None:
-    def fake_jwt_decode(*args: Any, **params: Any):
-        encoded = args[0]
-        if encoded == "fake id token":
-            return decoded_id_token
-        return userinfo
-
-    monkeypatch.setattr("jwt.decode", fake_jwt_decode)
-    monkeypatch.setattr("app.controllers.auth.env.FC_AMI_CLIENT_SECRET", "fake-client-secret")
-
-    NONCE = decoded_id_token["nonce"]
-    nonce = Nonce(nonce=NONCE)
-    db_session.add(nonce)
-    await db_session.commit()
-
-    fake_token_json_response = {
-        "access_token": "fake access token",
-        "expires_in": 60,
-        "id_token": "fake id token",
-        "scope": "openid given_name family_name preferred_username birthdate gender birthplace birthcountry email",
-        "token_type": "Bearer",
-    }
-    httpx_mock.add_response(
-        method="POST",
-        url="https://fcp-low.sbx.dev-franceconnect.fr/api/v2/token",
-        json=fake_token_json_response,
-        is_reusable=True,
-    )
-
-    auth = {"authorization": "Bearer fake access token"}
-    fake_userinfo_token = "fake userinfo jwt token"
-    httpx_mock.add_response(
-        method="GET",
-        url="https://fcp-low.sbx.dev-franceconnect.fr/api/v2/userinfo",
-        match_headers=auth,
-        text=fake_userinfo_token,
-        is_reusable=True,
-    )
-
-    address = {
-        "numero_libelle_voie": "1 RUE MONTORGUEIL",
-        "lieu_dit": "",
-        "code_postal_ville": "75002 PARIS",
-        "pays": "FRANCE",
-    }
-    fake_quotient_data = {"data": {"adresse": address}}
-    encoded_address = urlsafe_b64encode(json.dumps(address).encode("utf8")).decode("utf8")
-    httpx_mock.add_response(
-        method="GET",
-        url="https://staging.particulier.api.gouv.fr/v3/dss/quotient_familial/france_connect?recipient=13002526500013",
-        match_headers=auth,
-        json=fake_quotient_data,
-    )
-
-    response = test_client.get(
-        f"/login-callback?code=fake-code&state={nonce.id}", follow_redirects=False
-    )
-
-    assert response.status_code == 302
-    redirected_url = response.headers["location"]
-    assert redirected_url.startswith("https://localhost:5173")
-    assert url_contains_param(
-        "address",
-        encoded_address,
-        redirected_url,
-    )
-
-    nonce = Nonce(nonce=NONCE)
-    db_session.add(nonce)
-    await db_session.commit()
-    address = {
-        "numero_libelle_voie": "",
-        "lieu_dit": "",
-        "code_postal_ville": "",
-        "pays": "",
-    }
-    fake_quotient_data = {"data": {"adresse": address}}
-    encoded_address = urlsafe_b64encode(json.dumps(address).encode("utf8")).decode("utf8")
-    httpx_mock.add_response(
-        method="GET",
-        url="https://staging.particulier.api.gouv.fr/v3/dss/quotient_familial/france_connect?recipient=13002526500013",
-        match_headers=auth,
-        json=fake_quotient_data,
-    )
-
-    response = test_client.get(
-        f"/login-callback?code=fake-code&state={nonce.id}", follow_redirects=False
-    )
-
-    assert response.status_code == 302
-    redirected_url = response.headers["location"]
-    assert redirected_url.startswith("https://localhost:5173")
-    assert url_contains_param(
-        "address",
-        encoded_address,
-        redirected_url,
-    )
-
-
-async def test_login_callback_address_no_data(
-    test_client: TestClient[Litestar],
-    db_session: AsyncSession,
-    httpx_mock: HTTPXMock,
-    monkeypatch: pytest.MonkeyPatch,
-    userinfo: dict[str, Any],
-    decoded_id_token: dict[str, Any],
-) -> None:
-    def fake_jwt_decode(*args: Any, **params: Any):
-        encoded = args[0]
-        if encoded == "fake id token":
-            return decoded_id_token
-        return userinfo
-
-    monkeypatch.setattr("jwt.decode", fake_jwt_decode)
-    monkeypatch.setattr("app.controllers.auth.env.FC_AMI_CLIENT_SECRET", "fake-client-secret")
-
-    NONCE = decoded_id_token["nonce"]
-    nonce = Nonce(nonce=NONCE)
-    db_session.add(nonce)
-    await db_session.commit()
-
-    fake_token_json_response = {
-        "access_token": "fake access token",
-        "expires_in": 60,
-        "id_token": "fake id token",
-        "scope": "openid given_name family_name preferred_username birthdate gender birthplace birthcountry email",
-        "token_type": "Bearer",
-    }
-    httpx_mock.add_response(
-        method="POST",
-        url="https://fcp-low.sbx.dev-franceconnect.fr/api/v2/token",
-        json=fake_token_json_response,
-        is_reusable=True,
-    )
-
-    auth = {"authorization": "Bearer fake access token"}
-    fake_userinfo_token = "fake userinfo jwt token"
-    httpx_mock.add_response(
-        method="GET",
-        url="https://fcp-low.sbx.dev-franceconnect.fr/api/v2/userinfo",
-        match_headers=auth,
-        text=fake_userinfo_token,
-        is_reusable=True,
-    )
-
-    fake_quotient_datas: list[dict[str, Any]] = [
-        {},
-        {"data": {}},
-        {"data": {"address": {}}},
-    ]
-    for fake_quotient_data in fake_quotient_datas:
-        nonce = Nonce(nonce=NONCE)
-        db_session.add(nonce)
-        await db_session.commit()
-
-        httpx_mock.add_response(
-            method="GET",
-            url="https://staging.particulier.api.gouv.fr/v3/dss/quotient_familial/france_connect?recipient=13002526500013",
-            match_headers=auth,
-            json=fake_quotient_data,
-        )
-
-        response = test_client.get(
-            f"/login-callback?code=fake-code&state={nonce.id}", follow_redirects=False
-        )
-
-        assert response.status_code == 302
-        redirected_url = response.headers["location"]
-        assert redirected_url.startswith("https://localhost:5173")
-        assert "address" not in redirected_url
-
-
-async def test_login_callback_token_query_failure(
-    test_client: TestClient[Litestar],
-    db_session: AsyncSession,
-    httpx_mock: HTTPXMock,
-) -> None:
-    NONCE = "some random nonce"
-    nonce = Nonce(nonce=NONCE)
-    db_session.add(nonce)
-    await db_session.commit()
-
-    token_failure_response = {
-        "error": "invalid_grant",
-        "error_description": " grant request is invalid (authorization code not found)",
-        "error_uri": "https://docs.partenaires.franceconnect.gouv.fr/fs/fs-technique/fs-technique-erreurs/"
-        "?code=Y049E20B&id=801d508c-72d7-459d-8947-104cf89ce015",
-    }
-    httpx_mock.add_response(
-        method="POST",
-        url="https://fcp-low.sbx.dev-franceconnect.fr/api/v2/token",
-        json=token_failure_response,
-        status_code=401,
-    )
-
-    response = test_client.get(
-        f"/login-callback?code=fake-code&state={nonce.id}", follow_redirects=False
-    )
-
-    assert response.status_code == 302
-    redirected_url = response.headers["location"]
-    assert (
-        redirected_url
-        == "https://localhost:5173/?error=Erreur+lors+de+la+FranceConnexion%2C+veuillez+r%C3%A9essayer+plus+tard.&error_type=FranceConnect"
-    )
-
-
-async def test_login_callback_userinfo_query_failure(
-    test_client: TestClient[Litestar],
-    db_session: AsyncSession,
-    httpx_mock: HTTPXMock,
-    monkeypatch: pytest.MonkeyPatch,
-    decoded_id_token: dict[str, Any],
-) -> None:
-    def fake_jwt_decode(*args: Any, **params: Any):
-        return decoded_id_token
-
-    monkeypatch.setattr("jwt.decode", fake_jwt_decode)
-    monkeypatch.setattr("app.controllers.auth.env.FC_AMI_CLIENT_SECRET", "fake-client-secret")
-    monkeypatch.setattr(
-        "app.controllers.auth.env.PUBLIC_FC_SCOPE",
-        env.PUBLIC_FC_SCOPE.replace(" cnaf_quotient_familial", ""),
-    )
-
-    NONCE = decoded_id_token["nonce"]
-    nonce = Nonce(nonce=NONCE)
-    db_session.add(nonce)
-    await db_session.commit()
-
-    fake_token_json_response = {
-        "access_token": "fake access token",
-        "expires_in": 60,
-        "id_token": "fake id token",
-        "scope": "openid given_name family_name preferred_username birthdate gender birthplace birthcountry email",
-        "token_type": "Bearer",
-    }
-    httpx_mock.add_response(
-        method="POST",
-        url="https://fcp-low.sbx.dev-franceconnect.fr/api/v2/token",
-        json=fake_token_json_response,
-    )
-
-    auth = {"authorization": "Bearer fake access token"}
-    httpx_mock.add_response(
-        method="GET",
-        url="https://fcp-low.sbx.dev-franceconnect.fr/api/v2/userinfo",
-        match_headers=auth,
-        status_code=500,
-    )
-
-    response = test_client.get(
-        f"/login-callback?code=fake-code&state={nonce.id}", follow_redirects=False
-    )
-
-    assert response.status_code == 302
-    redirected_url = response.headers["location"]
-    assert (
-        redirected_url
-        == "https://localhost:5173/?error=Erreur+lors+de+la+FranceConnexion%2C+veuillez+r%C3%A9essayer+plus+tard.&error_type=FranceConnect"
-    )
-
-
-async def test_login_callback_address_query_failure(
-    test_client: TestClient[Litestar],
-    db_session: AsyncSession,
-    httpx_mock: HTTPXMock,
-    monkeypatch: pytest.MonkeyPatch,
-    userinfo: dict[str, Any],
-    decoded_id_token: dict[str, Any],
-) -> None:
-    def fake_jwt_decode(*args: Any, **params: Any):
-        encoded = args[0]
-        if encoded == "fake id token":
-            return decoded_id_token
-        return userinfo
-
-    monkeypatch.setattr("jwt.decode", fake_jwt_decode)
-    monkeypatch.setattr("app.controllers.auth.env.FC_AMI_CLIENT_SECRET", "fake-client-secret")
-
-    NONCE = decoded_id_token["nonce"]
-    nonce = Nonce(nonce=NONCE)
-    db_session.add(nonce)
-    await db_session.commit()
-
-    fake_token_json_response = {
-        "access_token": "fake access token",
-        "expires_in": 60,
-        "id_token": "fake id token",
-        "scope": "openid given_name family_name preferred_username birthdate gender birthplace birthcountry email",
-        "token_type": "Bearer",
-    }
-    httpx_mock.add_response(
-        method="POST",
-        url="https://fcp-low.sbx.dev-franceconnect.fr/api/v2/token",
-        json=fake_token_json_response,
-    )
-
-    auth = {"authorization": "Bearer fake access token"}
-    fake_userinfo_token = "fake userinfo jwt token"
-    httpx_mock.add_response(
-        method="GET",
-        url="https://fcp-low.sbx.dev-franceconnect.fr/api/v2/userinfo",
-        match_headers=auth,
-        text=fake_userinfo_token,
-    )
-
-    httpx_mock.add_response(
-        method="GET",
-        url="https://staging.particulier.api.gouv.fr/v3/dss/quotient_familial/france_connect?recipient=13002526500013",
-        match_headers=auth,
-        status_code=500,
-    )
-
-    response = test_client.get(
-        f"/login-callback?code=fake-code&state={nonce.id}", follow_redirects=False
-    )
-
-    assert response.status_code == 302
-    redirected_url = response.headers["location"]
-    assert redirected_url.startswith("https://localhost:5173")
-    assert "address" not in redirected_url
-
-
-async def test_login_callback_fc_error(
-    test_client: TestClient[Litestar],
-) -> None:
-    response = test_client.get(
-        "/login-callback?error=access_denied&error_description=User+auth+aborted&state=some-state",
-        follow_redirects=False,
-    )
-
-    assert response.status_code == 302
-    redirected_url = response.headers["location"]
-    assert url_contains_param("error_type", "FranceConnect", redirected_url)
-    assert url_contains_param("error", "access_denied", redirected_url)
-    assert url_contains_param("error_description", "User auth aborted", redirected_url)
-    assert url_contains_param("code", "fc_error", redirected_url)
-
-
-async def test_login_callback_error(
-    test_client: TestClient[Litestar],
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    def fake_retry():
-        raise Exception()
-
-    monkeypatch.setattr("app.controllers.auth.retry_fc_later", fake_retry)
-    response = test_client.get("/login-callback?state=some-state", follow_redirects=False)
-    redirected_url = response.headers["location"]
-    assert redirected_url == "https://localhost:5173/#/technical-error"
-
-
-async def test_logout(
-    user: User,
-    test_client: TestClient[Litestar],
-) -> None:
-    login(user, test_client)
-    assert test_client.cookies.get(jwt_cookie_auth.key)
-    response = test_client.post("/logout")
-    assert response.status_code == 201
-    assert not response.cookies.get(jwt_cookie_auth.key)
-
-
-async def test_logout_without_auth(
-    test_client: TestClient[Litestar],
-) -> None:
-    await assert_query_fails_without_auth("/logout", test_client, method="post")
-
-
-async def test_check_auth(
-    user: User,
-    test_client: TestClient[Litestar],
-) -> None:
-    login(user, test_client)
-    response = test_client.get("/check-auth")
-    assert response.status_code == 200
-    assert response.json() == {"authenticated": True}
-
-
-async def test_check_auth_without_auth(
-    test_client: TestClient[Litestar],
-) -> None:
-    await assert_query_fails_without_auth("/check-auth", test_client)
